Extract SurveyRetriesPanel to load histograms on demand
We don't want to load stats from the server if the histograms are
not to be shown.

See #1526
