+++ web/static/js/components/surveys/SurveyShow.jsx
@@ -12,7 +12,7 @@ import sum from 'lodash/sum'
 import { modeLabel } from '../../questionnaire.mode'
 import { referenceColorClasses, referenceColors, referenceColorClassForUnassigned, referenceColorForUnassigned } from '../../referenceColors'
 import classNames from 'classnames/bind'
-import { Stats, Forecasts, SuccessRate, QueueSize, RetriesHistogram } from '@instedd/surveda-d3-components'
+import { Stats, Forecasts, SuccessRate, QueueSize } from '@instedd/surveda-d3-components'
 import { translate } from 'react-i18next'
 
 type State = {
@@ -49,7 +49,6 @@ class SurveyShow extends Component<any, State> {
     pending: PropTypes.number,
     multiplier: PropTypes.number,
     needed: PropTypes.number,
-    retriesHistograms: PropTypes.array,
     overviewType: PropTypes.string
   }
 
@@ -65,7 +64,6 @@ class SurveyShow extends Component<any, State> {
     dispatch(actions.fetchSurveyIfNeeded(projectId, surveyId))
     dispatch(respondentActions.fetchRespondentsStats(projectId, surveyId))
     dispatch(actions.fetchSurveyStats(projectId, surveyId))
-    dispatch(actions.fetchSurveyRetriesHistograms(projectId, surveyId))
   }
 
   componentDidUpdate() {
@@ -134,95 +132,10 @@ class SurveyShow extends Component<any, State> {
     dispatch(actions.toggleLock())
   }
 
-  typeForRetriesHistogram(type) {
-    switch (type) {
-      case 'ivr':
-        return 'voice'
-      case 'end':
-        return 'discard'
-    }
-    return type
-  }
-
-  delayForRetriesHistogram(delay, isFirstAttempt) {
-    // It doesn't feel right hardcoding 1 when the delay is 0
-    // But we need this for when the delay is less than an hour
-    // Otherwise, the hole graph crashes
-    return isFirstAttempt ? delay : delay || 1
-  }
-
-  flowForRetriesHistogram(flow) {
-    const cloned = [...flow]
-    let offset = 0
-    cloned.forEach(step => {
-      offset += step.delay
-      step.offset = offset
-    })
-    return cloned.map(({delay, type, label, offset}, idx) => ({
-      delay: this.delayForRetriesHistogram(delay, idx == 0),
-      type: this.typeForRetriesHistogram(type),
-      label,
-      offset
-    }))
-  }
-
   render() {
     const { questionnaires, survey, respondentsByDisposition, reference, contactedRespondents, cumulativePercentages, target, project, t,
-      estimatedSuccessRate, initialSuccessRate, successRate, completionRate, completes, missing, pending, multiplier, needed, retriesHistograms, overviewType, router } = this.props
+      estimatedSuccessRate, initialSuccessRate, successRate, completionRate, completes, missing, pending, multiplier, needed, overviewType, router } = this.props
     const { stopUnderstood } = this.state
-
-    const getHistogram = h => {
-      const flow = this.flowForRetriesHistogram(h.flow)
-      const histLength = flow.reduce((total, attempt) => total + attempt.delay, 1)
-      const histActives = new Array(histLength)
-
-      let i
-      for (i = 0; i < histLength; i++) {
-        histActives[i] = {value: 0}
-      }
-      h.actives.forEach(slot => {
-        histActives[slot.hour].value = slot.respondents
-      })
-
-      return {
-        actives: histActives,
-        flow: flow,
-        // It doesn't apply to the current iteration
-        completes: histActives.map(() => false),
-        // It doesn't apply to the current iteration
-        timewindows: histActives.map(() => true)
-      }
-    }
-
-    const getHistograms = () => {
-      if (!retriesHistograms) return t('Loading...')
-      return retriesHistograms.map(h => getHistogram(h)).map(h => <RetriesHistogram
-        // Used in the y-axis height calculation
-        quota={target}
-        // Every attempt in the mode sequence, including the delay between them
-        flow={h.flow}
-        // Every set of active respondents grouped and ordered by hour
-        actives={h.actives}
-        // Every set of completed respondents grouped and ordered by hour
-        completes={h.completes}
-        // Every hour while the survey is active is flagged as true
-        timewindows={h.timewindows}
-        // Text about the current and next state of the survey regarding its schedule
-        // It doesn't apply to the current iteration
-        scheduleDescription=''
-        // Graph references
-        references={[
-          {
-            label: t('Trying'),
-            className: 'trying'
-          },
-          {
-            label: t('Stand by'),
-            className: 'standby'
-          }
-        ]}
-        />)
-    }
 
     if (!survey || !cumulativePercentages || !questionnaires || !respondentsByDisposition || !reference) {
       return <p>{t('Loading...')}</p>
@@ -383,19 +296,7 @@ class SurveyShow extends Component<any, State> {
               </div>
               <Stats data={stats} />
               <Forecasts data={forecasts} ceil={100} forecast={survey.state == 'running'} />
-              {
-                this.showHistograms()
-                ? (
-                  <div className='retries-histogram' style={{'marginTop': '20px'}}>
-                    <div className='header'>
-                      <div className='title'>{t('Retries histograms')}</div>
-                      <div className='description'>{t('Number of contacts in each stage of the retry schedule')}</div>
-                    </div>
-                    { getHistograms() }
-                  </div>
-                )
-                : null
-              }
+              { this.showHistograms() ? <SurveyRetriesPanel /> : null }
               <div className='row' style={{ 'display': 'flex', 'alignItems': 'center', 'marginTop': '20px' }}>
                 <div style={{ 'width': '50%' }}>
                   <div className='header'>
@@ -554,7 +455,6 @@ class SurveyShow extends Component<any, State> {
 const mapStateToProps = (state, ownProps) => {
   const respondentsStatsRoot = state.respondentsStats[ownProps.params.surveyId]
   const surveyStats = state.surveyStats.surveyId == ownProps.params.surveyId ? state.surveyStats.stats : null
-  const surveyRetriesHistograms = state.surveyRetriesHistograms.surveyId == ownProps.params.surveyId ? state.surveyRetriesHistograms.histograms : null
 
   let respondentsByDisposition = null
   let cumulativePercentages = {}
@@ -596,7 +496,6 @@ const mapStateToProps = (state, ownProps) => {
     missing: surveyStats ? surveyStats.missing : null,
     needed: surveyStats ? surveyStats.needed : null,
     pending: surveyStats ? surveyStats.pending : null,
-    retriesHistograms: surveyRetriesHistograms,
     overviewType: ownProps.params.overviewType
   })
 }
